@@ -1,38 +1,26 @@
-[tool.poetry]
-name = "passwault"
+[project]
+name = "passvault-project"
 version = "0.1.0"
-description = "An encoder/decoder of passwords to images using steganography"
-authors = ["Pedro Ribeiro <pedromr21@gmail.com>"]
+description = "Add your description here"
 readme = "README.md"
-package-mode = true
-packages = [{ include = "passwault"}]
+requires-python = ">=3.13"
+dependencies = [
+    "bcrypt>=4.3.0",
+    "cryptography>=45.0.6",
+    "numpy>=2.3.2",
+    "pillow>=11.3.0",
+    "pytest>=8.4.1",
+    "pytest-cov>=6.2.1",
+    "pytest-mock>=3.14.1",
+    "sqlcipher3>=0.5.4",
+]
 
-[tool.black]
-line-length = 150
-skip-string-normalization = true
-
-[tool.flake8]
-max-line-length = 150
-extend-ignore = ["E203", "W503", "E501"]
-exclude = ["venv", ".venv", "build", "dist"]
-
-[tool.poetry.dependencies]
-python = "^3.11"
-pillow = "^10.4.0"
-numpy = "^2.1.1"
-bcrypt = "^4.2.0"
-pytest = "^8.3.3"
-pytest-cov = "^6.0.0"
-pytest-mock = "^3.14.0"
-cryptography = "^44.0.2"
-pysqlcipher3 = "^1.2.0"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.3"
-pre-commit = "^3.8.0"
-flake8 = "^7.2.0"
-black = "^25.1.0"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[dependency-groups]
+dev = [
+    "black>=25.1.0",
+    "flake8>=7.3.0",
+    "pre-commit>=4.3.0",
+    "pytest>=8.4.1",
+    "pytest-cov>=6.2.1",
+    "pytest-mock>=3.14.1",
+]
