@@ -4,9 +4,9 @@ import json
 from pathlib import Path
 from typing import Tuple
 
-ALLOWED_EXT = ['.csv','.json']
+ALLOWED_EXT = [".csv", ".json"]
 
-VALID_IMAGE_EXTENSIONS = ['.jpg', '.png', '.gif', '.jpeg', '.tiff', '.bmp']
+VALID_IMAGE_EXTENSIONS = [".jpg", ".png", ".gif", ".jpeg", ".tiff", ".bmp"]
 
 
 def valid_image_file(file: str) -> bool:
@@ -16,7 +16,9 @@ def valid_image_file(file: str) -> bool:
         raise argparse.ArgumentTypeError("Invalid file path")
 
     if image_file_path.suffix not in VALID_IMAGE_EXTENSIONS:
-        raise argparse.ArgumentTypeError(f"File must have one of the following extension: {'.'.join(VALID_IMAGE_EXTENSIONS)}")
+        raise argparse.ArgumentTypeError(
+            f"File must have one of the following extension: {'.'.join(VALID_IMAGE_EXTENSIONS)}"
+        )
 
     return file
 
@@ -28,7 +30,9 @@ def valid_file(file: str) -> bool:
         raise argparse.ArgumentTypeError("Invalid file path")
 
     if file_path.suffix not in ALLOWED_EXT:
-        raise argparse.ArgumentTypeError(f"File must have one of the following extensions: {','.join(ALLOWED_EXT)}")
+        raise argparse.ArgumentTypeError(
+            f"File must have one of the following extensions: {','.join(ALLOWED_EXT)}"
+        )
 
     return file
 
@@ -36,7 +40,7 @@ def valid_file(file: str) -> bool:
 def _read_csv(file_path: Path):
     pw_pairs = []
     try:
-        with open(file_path, 'r', newline='') as csv_file:
+        with open(file_path, "r", newline="") as csv_file:
             sample = csv_file.read(2048)
             csv_file.seek(0)
             sniffer = csv.Sniffer()
@@ -57,34 +61,31 @@ def _read_csv(file_path: Path):
     except Exception as e:
         raise ValueError(f"Error with '{file_path.name}': {str(e)}")
 
+
 def _read_json(file_path: Path):
     try:
         result = []
         with open(file_path, "r") as f:
             data = json.load(f)
             for k, v in data.items():
-                if 'password' not in v or v['password'] == "":
-                    raise Exception('password is missing or empty.')
-                if 'username' not in v:
-                    raise Exception('the field password is missing.')
-                if v['username'] == '':
-                    v['username'] = None
-                
-                result.append((v['username'], v['password'], k))
-                
+                if "password" not in v or v["password"] == "":
+                    raise Exception("password is missing or empty.")
+                if "username" not in v:
+                    raise Exception("the field password is missing.")
+                if v["username"] == "":
+                    v["username"] = None
+
+                result.append((v["username"], v["password"], k))
+
         return result
-            
-            
+
     except Exception as e:
         raise ValueError(f"Error with '{file_path.name}': {str(e)}")
-
 
 
 def read_file(file: str) -> Tuple[Tuple[str, str]] | None:
     file_path = Path(file)
     file_suffix = file_path.suffix
-
-
 
     if file_suffix == ".csv":
         result = _read_csv(file_path)
@@ -93,4 +94,3 @@ def read_file(file: str) -> Tuple[Tuple[str, str]] | None:
     if file_suffix == ".json":
         result = _read_json(file_path)
         return result
-       
