@@ -2,7 +2,7 @@ import argparse
 from .embedder import Embedder
 from .utils import verifier
 
-VALID_IMAGE_EXTENSIONS = ['.jpg', '.png', '.gif', '.jpeg', '.tiff', '.bmp']
+VALID_IMAGE_EXTENSIONS = [".jpg", ".png", ".gif", ".jpeg", ".tiff", ".bmp"]
 
 
 def main():
@@ -10,7 +10,9 @@ def main():
         prog="ImagePass",
         description="Insert an image path and a password to encode the password into the image",
     )
-    parser.add_argument("option", choices=["encode", "decode"], help="the operation to be done")
+    parser.add_argument(
+        "option", choices=["encode", "decode"], help="the operation to be done"
+    )
     parser.add_argument("image_path", help="the image to encode or decode the password")
     parser.add_argument("-p", "--password", help="the password to be encoded")
     args = parser.parse_args()
@@ -19,14 +21,18 @@ def main():
         raise ValueError("ERROR: The path is not valid")
 
     if verifier.get_file_extension(args.image_path) not in VALID_IMAGE_EXTENSIONS:
-        raise ValueError(f"ERROR: not a valid image extension. VALID EXTENSIONS: {' '.join(VALID_IMAGE_EXTENSIONS)}")
+        raise ValueError(
+            f"ERROR: not a valid image extension. VALID EXTENSIONS: {' '.join(VALID_IMAGE_EXTENSIONS)}"
+        )
+
+    embedder = Embedder(image_path=args.image_path)
 
     if args.option == "encode":
         if not args.password:
-            raise ValueError("ERROR: You need to insert the password that will be encoded")
-        Embedder().encode(args.image_path, args.password)
+            raise ValueError("ERROR: You need to insert the password to be encoded")
+        embedder.encode(args.password)
     else:
-        Embedder().decode(args.image_path)
+        embedder.decode()
 
 
 if __name__ == "__main__":
